# pos-backend/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: POS Papelería — API REST
  version: 1.0.0
  description: |
    API del sistema de punto de venta para papelería.
    Desarrollada con Node.js/Express + Supabase (PostgreSQL).

    ## Autenticación
    Todos los endpoints (salvo `/auth/login`) requieren un **JWT Bearer token**
    obtenido desde `POST /api/auth/login`.

    ```
    Authorization: Bearer <token>
    ```

    ## Roles
    | Rol | Permisos |
    |---|---|
    | `cajero` | ventas, consulta de productos e inventario |
    | `administrador` | todo lo anterior + productos, inventario, corte de caja |

    ## Formato de respuesta exitosa
    ```json
    { "success": true, "data": { ... } }
    ```

    ## Formato de error
    ```json
    { "success": false, "error": "Mensaje descriptivo" }
    ```
servers:
  - url: http://localhost:3000/api
    description: Desarrollo local
  - url: http://192.168.1.100:3000/api
    description: Red local ITH

# ──────────────────────────────────────────────────────────────────────────────
# Security scheme
# ──────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido desde POST /auth/login. Expira en 10 horas.

  # ── Schemas reutilizables ────────────────────────────────────────────────────
  schemas:
...
    Error400:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Mensaje de validación" }
    Error401:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Token inválido o expirado" }
    Error403:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Se requiere rol administrador" }
    Error404:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Recurso no encontrado" }
...
    DetalleVenta:
      type: object
      properties:
        id_detalle:      { type: integer, example: 55 }
        id_venta:        { type: integer, example: 12 }
        id_producto:     { type: integer, example: 7 }
        cantidad:        { type: integer, example: 2 }
        precio_unitario: { type: number,  format: decimal, example: 35.00 }
        descuento_linea: { type: number,  format: decimal, example: 0 }
        subtotal_linea:  { type: number,  format: decimal, example: 70.00 }
        productos:
          type: object
          properties:
            nombre:        { type: string, example: "Cuaderno universitario 100 hojas" }
            codigo_barras: { type: string, example: "7501000000007" }
    Venta:
      type: object
      properties:
        id_venta:           { type: integer, example: 12 }
        folio:              { type: string,  example: "VTA-20260624-0012" }
        fecha_hora:         { type: string,  format: date-time }
        id_usuario:         { type: integer, example: 3 }
        subtotal:           { type: number,  format: decimal, example: 95.50 }
        descuento:          { type: number,  format: decimal, example: 0 }
        total:              { type: number,  format: decimal, example: 95.50 }
        metodo_pago:
          type: string
          enum: [efectivo, tarjeta, mixto]
          example: efectivo
        monto_recibido:     { type: number,  format: decimal, example: 100.00 }
        cambio:             { type: number,  format: decimal, example: 4.50 }
        estado:
          type: string
          enum: [completada, cancelada, devolucion]
          example: completada
        motivo_cancelacion: { type: string,  nullable: true }
        id_corte:           { type: integer, nullable: true }
        usuarios:
          type: object
          properties:
            nombre: { type: string, example: "María López" }
        detalle_venta:
          type: array
          items:
            $ref: '#/components/schemas/DetalleVenta'
...
    CorteCaja:
      type: object
      properties:
        id_corte:                { type: integer, example: 5 }
        fecha_corte:             { type: string,  format: date, example: "2026-06-24" }
        fecha_apertura:          { type: string,  format: date-time }
        fecha_cierre:            { type: string,  format: date-time, nullable: true }
        id_usuario:              { type: integer, example: 1 }
        total_ventas_sistema:    { type: number,  format: decimal, example: 3240.00 }
        total_efectivo_sistema:  { type: number,  format: decimal, example: 1850.00 }
        total_tarjeta_sistema:   { type: number,  format: decimal, example: 1390.00 }
        total_devoluciones:      { type: number,  format: decimal, example: 120.00 }
        monto_contado_fisico:    { type: number,  format: decimal, example: 1900.00, nullable: true }
        diferencia:
          type: number
          format: decimal
          example: 50.00
          nullable: true
          description: >
            Calculado automáticamente por PostgreSQL como
            (monto_contado_fisico − total_efectivo_sistema). Campo GENERATED ALWAYS AS.
        observaciones:           { type: string,  nullable: true }
        estado:
          type: string
          enum: [abierto, cerrado]
          example: cerrado
        num_ventas:
          type: integer
          example: 28
          description: Ventas completadas. Calculado por el service, no persistido.
        num_canceladas:
          type: integer
          example: 2
          description: Ventas canceladas. Calculado por el service, no persistido.
        usuarios:
          type: object
          properties:
            nombre: { type: string, example: "Administrador" }

# ──────────────────────────────────────────────────────────────────────────────
# Security global (todos los endpoints salvo /auth/login)
# ──────────────────────────────────────────────────────────────────────────────
security:
  - bearerAuth: []

# ──────────────────────────────────────────────────────────────────────────────
# Tags
# ──────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /corte-caja:
    post:
      tags: [Corte de Caja]
      summary: Abrir corte del día
      description: |
        Abre el corte de caja para la fecha actual.
        Solo se permite **un corte por día** (`fecha_corte UNIQUE`).

        **Requiere rol administrador.**

        No acepta body — la fecha y el usuario se toman del token JWT y del sistema.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              description: Body vacío (no se requiere ningún campo)
      responses:
        '201':
          description: Corte abierto
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:    { $ref: '#/components/schemas/CorteCaja' }
              example:
                success: true
                data:
                  id_corte: 5
                  fecha_corte: "2026-06-24"
                  fecha_apertura: "2026-06-24T08:00:00.000Z"
                  id_usuario: 1
                  estado: abierto
                  total_ventas_sistema: 0
        '400':
          description: Ya existe un corte para hoy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error400' }
              example: { success: false, error: "Ya existe un corte para hoy" }
        '401': { description: No autenticado,   content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }
        '403': { description: Rol insuficiente, content: { application/json: { schema: { $ref: '#/components/schemas/Error403' } } } }

    get:
      tags: [Corte de Caja]
      summary: Historial de cortes
      description: |
        Lista todos los cortes de caja con filtros opcionales de fecha.
        **Requiere rol administrador.**
      parameters:
        - name: desde
          in: query
          description: Fecha inicio (YYYY-MM-DD) — filtra por fecha_apertura
          schema: { type: string, format: date, example: "2026-06-01" }
        - name: hasta
          in: query
          description: Fecha fin (YYYY-MM-DD) — se extiende hasta las 23:59:59
          schema: { type: string, format: date, example: "2026-06-30" }
      responses:
        '200':
          description: Lista de cortes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CorteCaja' }
        '401': { description: No autenticado,   content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }
        '403': { description: Rol insuficiente, content: { application/json: { schema: { $ref: '#/components/schemas/Error403' } } } }
  /corte-caja/{id}:
    get:
      tags: [Corte de Caja]
      summary: Obtener corte por ID
      description: |
        Devuelve el corte con las ventas asociadas (aquellas cuyo `id_corte` coincide).
        **Requiere rol administrador.**
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 5 }
      responses:
        '200':
          description: Corte con ventas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    allOf:
                      - $ref: '#/components/schemas/CorteCaja'
                      - type: object
                        properties:
                          ventas:
                            type: array
                            items: { $ref: '#/components/schemas/Venta' }
        '404': { description: Corte no encontrado, content: { application/json: { schema: { $ref: '#/components/schemas/Error404' } } } }
        '401': { description: No autenticado,      content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }
        '403': { description: Rol insuficiente,    content: { application/json: { schema: { $ref: '#/components/schemas/Error403' } } } }
  /corte-caja/{id}/cerrar:
    put:
      tags: [Corte de Caja]
      summary: Cerrar corte y descargar PDF
      description: |
        Cierra el corte de caja activo y devuelve el reporte en **PDF** como descarga.

        **Requiere rol administrador.**

        **Proceso interno:**
        1. Verifica que el corte existe y está `abierto`
        2. Consulta todas las ventas del período
        3. Calcula totales por método de pago y estado
        4. Actualiza `ventas.id_corte` en lote
        5. Actualiza `corte_caja` con los totales calculados
        6. PostgreSQL recalcula `diferencia` automáticamente (columna GENERATED)
        7. Genera y envía el PDF del reporte

        > **Nota:** `diferencia = monto_contado_fisico − total_efectivo_sistema`
        > es calculado por PostgreSQL. El backend nunca la incluye en el UPDATE.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 5 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [monto_contado_fisico]
              properties:
                monto_contado_fisico:
                  type: number
                  minimum: 0
                  example: 1900.00
                  description: Efectivo contado físicamente al cerrar el turno
                observaciones:
                  type: string
                  example: "Cierre de turno matutino. Sin incidencias."
            examples:
              cierre_normal:
                summary: Cierre sin diferencias
                value:
                  monto_contado_fisico: 1850.00
                  observaciones: "Cierre de turno"
              cierre_con_faltante:
                summary: Cierre con faltante de $50
                value:
                  monto_contado_fisico: 1800.00
                  observaciones: "Faltante de $50 — revisar ticket 0041"
      responses:
        '200':
          description: Reporte PDF del corte cerrado
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="corte-5.pdf"'
        '400':
          description: Error de negocio
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error400' }
              examples:
                ya_cerrado:
                  value: { success: false, error: "El corte ya está cerrado" }
                monto_negativo:
                  value: { success: false, error: "El monto contado no puede ser negativo" }
        '401': { description: No autenticado,     content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }
        '403': { description: Rol insuficiente,   content: { application/json: { schema: { $ref: '#/components/schemas/Error403' } } } }
        '404': { description: Corte no encontrado, content: { application/json: { schema: { $ref: '#/components/schemas/Error404' } } } }